fix(auth): correct admin access logic and add route guards
Ensure that only authorized admins can access admin-specific routes by fixing the conditional logic in MainHeader.jsx. Added route guards using useRouteIfAuthorizedAndHeIsNotAdmin and useRouteIfAuthorizedAndHeIsAdmin hooks to prevent unauthorized access to admin and user pages respectively.
