+++ client/src/util/useRouteIfNotAuthorized.js
@@ -10,3 +10,31 @@ export function useRouteIfNotAuthorized() {
     }
   }, [isAuthorized, navigate]);
 }
+export function useRouteIfAuthorizedAndHeIsAdmin() {
+  const navigate = useNavigate();
+  const { isAuthorized, user } = useAuth();
+  useEffect(() => {
+    if (!isAuthorized) {
+      navigate("/");
+      return;
+    }
+    if (isAuthorized && user.isAdmin) {
+      navigate("/admin/admin-home");
+      return;
+    }
+  }, [isAuthorized, user?.isAdmin, navigate]);
+}
+export function useRouteIfAuthorizedAndHeIsNotAdmin() {
+    const navigate = useNavigate();
+    const { isAuthorized, user } = useAuth();
+    useEffect(() => {
+      if (!isAuthorized) {
+        navigate("/");
+        return;
+      }
+      if (isAuthorized && !user.isAdmin) {
+        navigate("/home");
+        return;
+      }
+    }, [isAuthorized, user?.isAdmin, navigate]);
+  }
